+++ OTAcceleratorCore/OTMultiPartyCommunicator.m
@@ -133,6 +133,8 @@
 @interface OTMultiPartyCommunicator() <OTSessionDelegate, OTSubscriberKitDelegate, OTPublisherDelegate, OTVideoViewProtocol>
 @property (nonatomic) BOOL isCallEnabled;
 @property (nonatomic) NSString *name;
+@property (nonatomic) NSUInteger connectionCount;
+@property (nonatomic) NSUInteger connectionCountOlderThanMe;
 @property (nonatomic) OTPublisher *publisher;
 @property (nonatomic) NSMutableArray *subscribers;
 @property (weak, nonatomic) OTAcceleratorSession *session;
@@ -270,6 +272,7 @@
     }
     
     self.isCallEnabled = NO;
+    _connectionCount = 0;
     return disconnectError;
 }
 
@@ -299,17 +302,20 @@
     [[MultiPartyLoggingWrapper sharedInstance].logger setSessionId:session.sessionId
                                                       connectionId:session.connection.connectionId
                                                          partnerId:@([self.session.apiKey integerValue])];
+    _connectionCount++;
     
     if (!self.publisher) {
         if (!self.screenSharingView) {
-            self.publisher = [[OTPublisher alloc] initWithDelegate:self
-                                                              name:self.name];
+            OTPublisherSettings *setting = [[OTPublisherSettings alloc] init];
+            setting.name = self.name;
+            self.publisher = [[OTPublisher alloc] initWithDelegate:self settings:setting];
         }
         else {
-            self.publisher = [[OTPublisher alloc] initWithDelegate:self
-                                                              name:self.name
-                                                        audioTrack:YES
-                                                        videoTrack:YES];
+            OTPublisherSettings *setting = [[OTPublisherSettings alloc] init];
+            setting.name = self.name;
+            setting.audioTrack = YES;
+            setting.videoTrack = YES;
+            self.publisher = [[OTPublisher alloc] initWithDelegate:self settings:setting];
             
             [self.publisher setVideoType:OTPublisherKitVideoTypeScreen];
             self.publisher.audioFallbackEnabled = NO;
@@ -383,10 +389,27 @@
     }
 }
 
+- (void)session:(OTSession*) session
+connectionCreated:(OTConnection*) connection {
+    _connectionCount++;
+    
+    //check creationtime of the connections
+    [self compareConnectionTimeWithConnection: connection];
+}
+
+- (void)session:(OTSession*) session
+connectionDestroyed:(OTConnection*) connection {
+    _connectionCount--;
+    
+    //check creationtime of the connections
+    [self compareConnectionTimeWithConnection: connection];
+}
+
 - (void)sessionDidDisconnect:(OTSession *)session {
     [self notifyAllWithSignal:OTPublisherDestroyed
                    subscriber:nil
                         error:nil];
+    _connectionCount = 0;
 }
 
 - (void)session:(OTSession *)session didFailWithError:(OTError *)error {
@@ -603,4 +626,15 @@
     }
 }
 
+- (BOOL)isFirstConnection {
+    return _connectionCountOlderThanMe > 0 ? false : true;
+}
+
+#pragma mark - Private Methods
+-(void)compareConnectionTimeWithConnection: (OTConnection *)connection {
+    if (self.session.connection) {
+        NSComparisonResult result = [connection.creationTime compare:_session.connection.creationTime];
+        result == NSOrderedDescending ? _connectionCountOlderThanMe++ : _connectionCountOlderThanMe--;
+    }
+}
 @end